[FIX] api view cbv & delete error fix

+++ api/views_FBV.py
@@ -1,22 +1,25 @@
-# views.py
+# views_FBV.py
 
-from rest_framework.views import APIView
-from rest_framework.generics import get_object_or_404
-from api.models import ToDo
-from api.serializers import ToDoSerializer
 from django.http import JsonResponse
 from rest_framework.parsers import JSONParser
-from rest_framework.response import Response
+from api.models import ToDo
+from api.serializers import ToDoSerializer
+from django.shortcuts import get_object_or_404
+from django.views.decorators.csrf import csrf_exempt
+from django.utils.decorators import method_decorator
 
 
-class TodoList(APIView):
-
-    def get(self, request):  # 모든 데이터 얻기
+@csrf_exempt
+def todo_list(request):
+    """
+    List all code snippets, or create a new snippet.
+    """
+    if request.method == 'GET':  # 모든 데이터 얻기
         todo = ToDo.objects.all()
         serializer = ToDoSerializer(todo, many=True)
         return JsonResponse(serializer.data, status=200, safe=False)
 
-    def post(self, request):  # 새로운 데이터 등록
+    elif request.method == 'POST':  # 새로운 데이터 등록
         data = JSONParser().parse(request)
         serializer = ToDoSerializer(data=data)
         if serializer.is_valid():
@@ -25,27 +28,23 @@ class TodoList(APIView):
         return JsonResponse(serializer.errors, status=400)
 
 
-class TodoDetail(APIView):
-
-    def get_object(self, id):
+@method_decorator(csrf_exempt, name='dispatch')
+def todo_one(request, id):
+    if request.method == 'GET':  # 특정 데이터 얻기
         todo = get_object_or_404(ToDo, id=id)
-        return todo
-
-    def get(self, request, id):  # 특정 데이터 얻기
-        todo = self.get_object(id)
         serializer = ToDoSerializer(todo)
         return JsonResponse(serializer.data, status=200, safe=False)
 
-    def delete(self, request, id):  # 특정 데이터 삭제
-        todo = self.get_object(id)
+    elif request.method == 'DELETE':  # 특정 데이터 삭제
+        todo = get_object_or_404(ToDo, id=id)
         todo.delete()
-        return Response(status=204)
+        return JsonResponse(status=204)
 
-    def patch(self, request, id):  # 특정 데이터 업데이트
-        todo = self.get_object(id)
+    elif request.method == 'PATCH':  # 특정 데이터 업데이트
+        todo = get_object_or_404(ToDo, id=id)
         data = JSONParser().parse(request)
         serializer = ToDoSerializer(todo, data=data)
         if serializer.is_valid():
             serializer.save()
             return JsonResponse(serializer.data, status=200)
-        return JsonResponse(serializer.errors, status=400)
+        return JsonResponse(status=400)
